# whatsudow_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsudow",
        "description": "Aplicación que permite a los organizadores de eventos unificar los datos de sus proveedores de servicios. Con esta app, los organizadores de eventos pueden almacenar y acceder a información importante de los proveedores, comocalendarios de disponibilidad, PDF actualizados, cláusulas y datos de contacto.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/availability": {
            "get": {
                "tags": [
                    "Availability"
                ],
                "summary": "Obtiene las disponibilidades del proveedor",
                "description": "Recupera las disponibilidades del proveedor.",
                "operationId": "getAvailabilities",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "disponible",
                                                    "pre-reservado",
                                                    "no-disponible"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Availability"
                ],
                "summary": "Crea una nueva disponibilidad",
                "description": "Crea una nueva disponibilidad para el proveedor y la almacena en la base de datos. También crea un evento en Google Calendar para esa disponibilidad.",
                "operationId": "createAvailability",
                "requestBody": {
                    "description": "Datos de disponibilidad",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "pre-reservado",
                                            "no-disponible"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilidad creada con éxito"
                    },
                    "403": {
                        "description": "No tienes permiso para realizar esta acción"
                    }
                }
            }
        },
...
        "/availability/events": {
            "get": {
                "tags": [
                    "Availability"
                ],
                "summary": "Obtener los eventos de disponibilidad",
                "description": "Recupera los eventos de disponibilidad para los proveedores. Cada evento está codificado por colores según el estado de disponibilidad: verde para disponible, amarillo para reserva previa y rojo para no disponible.",
                "operationId": "getAvailabilityEvents",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "start": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "end": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "color": {
                                                "type": "string"
                                            },
                                            "proveedor": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permiso para realizar esta acción",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}